+++ crates/ffmpeg/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed windows build (#238)
+
 ## [0.1.2](https://github.com/ScuffleCloud/scuffle/compare/scuffle-ffmpeg-v0.1.1...scuffle-ffmpeg-v0.1.2) - 2025-02-10
 
 ## [0.1.1](https://github.com/ScuffleCloud/scuffle/compare/scuffle-ffmpeg-v0.1.0...scuffle-ffmpeg-v0.1.1) - 2025-02-10